requirements.txt
azure-storage-blob
fastapi
matplotlib
openai
pytest
uvicorn
